WIP Refactor printer logic into states and subsystem objects.
