+++ octoprint_bambu_printer/printer/states/idle_state.py
@@ -1,0 +1,7 @@
+from __future__ import annotations
+
+from octoprint_bambu_printer.printer.states.a_printer_state import APrinterState
+
+
+class IdleState(APrinterState):
+    pass
